Refactor: Split Game Center types and tests into smaller files
- Split internal/asc/game_center.go into separate files:
  - game_center_achievements.go (achievement types)
  - game_center_leaderboards.go (leaderboard types)
  - game_center_leaderboard_sets.go (leaderboard set types)
- Split cmd/game_center_test.go into separate test files:
  - game_center_achievements_test.go
  - game_center_leaderboards_test.go
  - game_center_leaderboard_sets_test.go
