+++ cmd/game_center_achievements_test.go
@@ -175,169 +175,6 @@ func TestGameCenterAchievementsDeleteValidationErrors(t *testing.T) {
 	}
 }
 
-func TestGameCenterLeaderboardsListValidationErrors(t *testing.T) {
-	t.Setenv("ASC_APP_ID", "")
-
-	root := RootCommand("1.2.3")
-	root.FlagSet.SetOutput(io.Discard)
-
-	stdout, stderr := captureOutput(t, func() {
-		if err := root.Parse([]string{"game-center", "leaderboards", "list"}); err != nil {
-			t.Fatalf("parse error: %v", err)
-		}
-		err := root.Run(context.Background())
-		if !errors.Is(err, flag.ErrHelp) {
-			t.Fatalf("expected ErrHelp, got %v", err)
-		}
-	})
-
-	if stdout != "" {
-		t.Fatalf("expected empty stdout, got %q", stdout)
-	}
-	if !strings.Contains(stderr, "--app is required") {
-		t.Fatalf("expected missing app error, got %q", stderr)
-	}
-}
-
-func TestGameCenterLeaderboardsCreateValidationErrors(t *testing.T) {
-	t.Setenv("ASC_APP_ID", "")
-
-	tests := []struct {
-		name    string
-		args    []string
-		wantErr string
-	}{
-		{
-			name:    "missing app",
-			args:    []string{"game-center", "leaderboards", "create", "--reference-name", "Test", "--vendor-id", "com.test", "--formatter", "INTEGER", "--sort", "DESC", "--submission-type", "BEST_SCORE"},
-			wantErr: "--app is required",
-		},
-		{
-			name:    "missing reference-name",
-			args:    []string{"game-center", "leaderboards", "create", "--app", "APP_ID", "--vendor-id", "com.test", "--formatter", "INTEGER", "--sort", "DESC", "--submission-type", "BEST_SCORE"},
-			wantErr: "--reference-name is required",
-		},
-		{
-			name:    "missing vendor-id",
-			args:    []string{"game-center", "leaderboards", "create", "--app", "APP_ID", "--reference-name", "Test", "--formatter", "INTEGER", "--sort", "DESC", "--submission-type", "BEST_SCORE"},
-			wantErr: "--vendor-id is required",
-		},
-		{
-			name:    "missing formatter",
-			args:    []string{"game-center", "leaderboards", "create", "--app", "APP_ID", "--reference-name", "Test", "--vendor-id", "com.test", "--sort", "DESC", "--submission-type", "BEST_SCORE"},
-			wantErr: "--formatter is required",
-		},
-		{
-			name:    "missing sort",
-			args:    []string{"game-center", "leaderboards", "create", "--app", "APP_ID", "--reference-name", "Test", "--vendor-id", "com.test", "--formatter", "INTEGER", "--submission-type", "BEST_SCORE"},
-			wantErr: "--sort is required",
-		},
-		{
-			name:    "missing submission-type",
-			args:    []string{"game-center", "leaderboards", "create", "--app", "APP_ID", "--reference-name", "Test", "--vendor-id", "com.test", "--formatter", "INTEGER", "--sort", "DESC"},
-			wantErr: "--submission-type is required",
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			root := RootCommand("1.2.3")
-			root.FlagSet.SetOutput(io.Discard)
-
-			stdout, stderr := captureOutput(t, func() {
-				if err := root.Parse(test.args); err != nil {
-					t.Fatalf("parse error: %v", err)
-				}
-				err := root.Run(context.Background())
-				if !errors.Is(err, flag.ErrHelp) {
-					t.Fatalf("expected ErrHelp, got %v", err)
-				}
-			})
-
-			if stdout != "" {
-				t.Fatalf("expected empty stdout, got %q", stdout)
-			}
-			if !strings.Contains(stderr, test.wantErr) {
-				t.Fatalf("expected error %q, got %q", test.wantErr, stderr)
-			}
-		})
-	}
-}
-
-func TestGameCenterLeaderboardSetsListValidationErrors(t *testing.T) {
-	t.Setenv("ASC_APP_ID", "")
-
-	root := RootCommand("1.2.3")
-	root.FlagSet.SetOutput(io.Discard)
-
-	stdout, stderr := captureOutput(t, func() {
-		if err := root.Parse([]string{"game-center", "leaderboard-sets", "list"}); err != nil {
-			t.Fatalf("parse error: %v", err)
-		}
-		err := root.Run(context.Background())
-		if !errors.Is(err, flag.ErrHelp) {
-			t.Fatalf("expected ErrHelp, got %v", err)
-		}
-	})
-
-	if stdout != "" {
-		t.Fatalf("expected empty stdout, got %q", stdout)
-	}
-	if !strings.Contains(stderr, "--app is required") {
-		t.Fatalf("expected missing app error, got %q", stderr)
-	}
-}
-
-func TestGameCenterLeaderboardSetsCreateValidationErrors(t *testing.T) {
-	t.Setenv("ASC_APP_ID", "")
-
-	tests := []struct {
-		name    string
-		args    []string
-		wantErr string
-	}{
-		{
-			name:    "missing app",
-			args:    []string{"game-center", "leaderboard-sets", "create", "--reference-name", "Test", "--vendor-id", "com.test"},
-			wantErr: "--app is required",
-		},
-		{
-			name:    "missing reference-name",
-			args:    []string{"game-center", "leaderboard-sets", "create", "--app", "APP_ID", "--vendor-id", "com.test"},
-			wantErr: "--reference-name is required",
-		},
-		{
-			name:    "missing vendor-id",
-			args:    []string{"game-center", "leaderboard-sets", "create", "--app", "APP_ID", "--reference-name", "Test"},
-			wantErr: "--vendor-id is required",
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			root := RootCommand("1.2.3")
-			root.FlagSet.SetOutput(io.Discard)
-
-			stdout, stderr := captureOutput(t, func() {
-				if err := root.Parse(test.args); err != nil {
-					t.Fatalf("parse error: %v", err)
-				}
-				err := root.Run(context.Background())
-				if !errors.Is(err, flag.ErrHelp) {
-					t.Fatalf("expected ErrHelp, got %v", err)
-				}
-			})
-
-			if stdout != "" {
-				t.Fatalf("expected empty stdout, got %q", stdout)
-			}
-			if !strings.Contains(stderr, test.wantErr) {
-				t.Fatalf("expected error %q, got %q", test.wantErr, stderr)
-			}
-		})
-	}
-}
-
 func TestGameCenterAchievementLocalizationsListValidationErrors(t *testing.T) {
 	root := RootCommand("1.2.3")
 	root.FlagSet.SetOutput(io.Discard)
@@ -379,76 +216,6 @@ func TestGameCenterAchievementLocalizationsCreateValidationErrors(t *testing.T) 
 		{
 			name:    "missing name",
 			args:    []string{"game-center", "achievements", "localizations", "create", "--achievement-id", "ACH_ID", "--locale", "en-US", "--before-earned-description", "Before", "--after-earned-description", "After"},
-			wantErr: "--name is required",
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			root := RootCommand("1.2.3")
-			root.FlagSet.SetOutput(io.Discard)
-
-			stdout, stderr := captureOutput(t, func() {
-				if err := root.Parse(test.args); err != nil {
-					t.Fatalf("parse error: %v", err)
-				}
-				err := root.Run(context.Background())
-				if !errors.Is(err, flag.ErrHelp) {
-					t.Fatalf("expected ErrHelp, got %v", err)
-				}
-			})
-
-			if stdout != "" {
-				t.Fatalf("expected empty stdout, got %q", stdout)
-			}
-			if !strings.Contains(stderr, test.wantErr) {
-				t.Fatalf("expected error %q, got %q", test.wantErr, stderr)
-			}
-		})
-	}
-}
-
-func TestGameCenterLeaderboardLocalizationsListValidationErrors(t *testing.T) {
-	root := RootCommand("1.2.3")
-	root.FlagSet.SetOutput(io.Discard)
-
-	stdout, stderr := captureOutput(t, func() {
-		if err := root.Parse([]string{"game-center", "leaderboards", "localizations", "list"}); err != nil {
-			t.Fatalf("parse error: %v", err)
-		}
-		err := root.Run(context.Background())
-		if !errors.Is(err, flag.ErrHelp) {
-			t.Fatalf("expected ErrHelp, got %v", err)
-		}
-	})
-
-	if stdout != "" {
-		t.Fatalf("expected empty stdout, got %q", stdout)
-	}
-	if !strings.Contains(stderr, "--leaderboard-id is required") {
-		t.Fatalf("expected missing leaderboard-id error, got %q", stderr)
-	}
-}
-
-func TestGameCenterLeaderboardLocalizationsCreateValidationErrors(t *testing.T) {
-	tests := []struct {
-		name    string
-		args    []string
-		wantErr string
-	}{
-		{
-			name:    "missing leaderboard-id",
-			args:    []string{"game-center", "leaderboards", "localizations", "create", "--locale", "en-US", "--name", "Test"},
-			wantErr: "--leaderboard-id is required",
-		},
-		{
-			name:    "missing locale",
-			args:    []string{"game-center", "leaderboards", "localizations", "create", "--leaderboard-id", "LB_ID", "--name", "Test"},
-			wantErr: "--locale is required",
-		},
-		{
-			name:    "missing name",
-			args:    []string{"game-center", "leaderboards", "localizations", "create", "--leaderboard-id", "LB_ID", "--locale", "en-US"},
 			wantErr: "--name is required",
 		},
 	}
@@ -543,74 +310,26 @@ func TestGameCenterAchievementReleasesListValidationErrors(t *testing.T) {
 	}
 }
 
-func TestGameCenterLeaderboardSetMembersListValidationErrors(t *testing.T) {
+func TestGameCenterAchievementsListLimitValidation(t *testing.T) {
+	t.Setenv("ASC_APP_ID", "APP_ID")
+
 	root := RootCommand("1.2.3")
 	root.FlagSet.SetOutput(io.Discard)
 
-	stdout, stderr := captureOutput(t, func() {
-		if err := root.Parse([]string{"game-center", "leaderboard-sets", "members", "list"}); err != nil {
+	stdout, _ := captureOutput(t, func() {
+		if err := root.Parse([]string{"game-center", "achievements", "list", "--app", "APP_ID", "--limit", "201"}); err != nil {
 			t.Fatalf("parse error: %v", err)
 		}
 		err := root.Run(context.Background())
-		if !errors.Is(err, flag.ErrHelp) {
-			t.Fatalf("expected ErrHelp, got %v", err)
+		if err == nil {
+			t.Fatalf("expected error, got nil")
+		}
+		if !strings.Contains(err.Error(), "--limit must be between 1 and 200") {
+			t.Fatalf("expected error containing %q, got %v", "--limit must be between 1 and 200", err)
 		}
 	})
 
 	if stdout != "" {
 		t.Fatalf("expected empty stdout, got %q", stdout)
-	}
-	if !strings.Contains(stderr, "--set-id is required") {
-		t.Fatalf("expected missing set-id error, got %q", stderr)
-	}
-}
-
-func TestGameCenterLimitValidation(t *testing.T) {
-	t.Setenv("ASC_APP_ID", "APP_ID")
-
-	tests := []struct {
-		name    string
-		args    []string
-		wantErr string
-	}{
-		{
-			name:    "achievements list limit too high",
-			args:    []string{"game-center", "achievements", "list", "--app", "APP_ID", "--limit", "201"},
-			wantErr: "--limit must be between 1 and 200",
-		},
-		{
-			name:    "leaderboards list limit too high",
-			args:    []string{"game-center", "leaderboards", "list", "--app", "APP_ID", "--limit", "300"},
-			wantErr: "--limit must be between 1 and 200",
-		},
-		{
-			name:    "leaderboard-sets list limit too high",
-			args:    []string{"game-center", "leaderboard-sets", "list", "--app", "APP_ID", "--limit", "500"},
-			wantErr: "--limit must be between 1 and 200",
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			root := RootCommand("1.2.3")
-			root.FlagSet.SetOutput(io.Discard)
-
-			stdout, _ := captureOutput(t, func() {
-				if err := root.Parse(test.args); err != nil {
-					t.Fatalf("parse error: %v", err)
-				}
-				err := root.Run(context.Background())
-				if err == nil {
-					t.Fatalf("expected error, got nil")
-				}
-				if !strings.Contains(err.Error(), test.wantErr) {
-					t.Fatalf("expected error containing %q, got %v", test.wantErr, err)
-				}
-			})
-
-			if stdout != "" {
-				t.Fatalf("expected empty stdout, got %q", stdout)
-			}
-		})
 	}
 }
