+++ internal/asc/game_center_leaderboard_sets.go
@@ -1,0 +1,361 @@
+package asc
+
+import (
+	"net/url"
+	"strings"
+)
+
+// GameCenterLeaderboardSetAttributes represents a Game Center leaderboard set resource.
+type GameCenterLeaderboardSetAttributes struct {
+	ReferenceName    string `json:"referenceName"`
+	VendorIdentifier string `json:"vendorIdentifier"`
+}
+
+// GameCenterLeaderboardSetCreateAttributes describes attributes for creating a leaderboard set.
+type GameCenterLeaderboardSetCreateAttributes struct {
+	ReferenceName    string `json:"referenceName"`
+	VendorIdentifier string `json:"vendorIdentifier"`
+}
+
+// GameCenterLeaderboardSetUpdateAttributes describes attributes for updating a leaderboard set.
+type GameCenterLeaderboardSetUpdateAttributes struct {
+	ReferenceName *string `json:"referenceName,omitempty"`
+}
+
+// GameCenterLeaderboardSetRelationships describes relationships for leaderboard sets.
+type GameCenterLeaderboardSetRelationships struct {
+	GameCenterDetail *Relationship `json:"gameCenterDetail"`
+}
+
+// GameCenterLeaderboardSetCreateData is the data portion of a leaderboard set create request.
+type GameCenterLeaderboardSetCreateData struct {
+	Type          ResourceType                             `json:"type"`
+	Attributes    GameCenterLeaderboardSetCreateAttributes `json:"attributes"`
+	Relationships *GameCenterLeaderboardSetRelationships   `json:"relationships,omitempty"`
+}
+
+// GameCenterLeaderboardSetCreateRequest is a request to create a leaderboard set.
+type GameCenterLeaderboardSetCreateRequest struct {
+	Data GameCenterLeaderboardSetCreateData `json:"data"`
+}
+
+// GameCenterLeaderboardSetUpdateData is the data portion of a leaderboard set update request.
+type GameCenterLeaderboardSetUpdateData struct {
+	Type       ResourceType                              `json:"type"`
+	ID         string                                    `json:"id"`
+	Attributes *GameCenterLeaderboardSetUpdateAttributes `json:"attributes,omitempty"`
+}
+
+// GameCenterLeaderboardSetUpdateRequest is a request to update a leaderboard set.
+type GameCenterLeaderboardSetUpdateRequest struct {
+	Data GameCenterLeaderboardSetUpdateData `json:"data"`
+}
+
+// GameCenterLeaderboardSetsResponse is the response from leaderboard set list endpoints.
+type GameCenterLeaderboardSetsResponse = Response[GameCenterLeaderboardSetAttributes]
+
+// GameCenterLeaderboardSetResponse is the response from leaderboard set detail endpoints.
+type GameCenterLeaderboardSetResponse = SingleResponse[GameCenterLeaderboardSetAttributes]
+
+// GameCenterLeaderboardSetDeleteResult represents CLI output for leaderboard set deletions.
+type GameCenterLeaderboardSetDeleteResult struct {
+	ID      string `json:"id"`
+	Deleted bool   `json:"deleted"`
+}
+
+// GCLeaderboardSetsOption is a functional option for GetGameCenterLeaderboardSets.
+type GCLeaderboardSetsOption func(*gcLeaderboardSetsQuery)
+
+type gcLeaderboardSetsQuery struct {
+	listQuery
+}
+
+// WithGCLeaderboardSetsLimit sets the max number of leaderboard sets to return.
+func WithGCLeaderboardSetsLimit(limit int) GCLeaderboardSetsOption {
+	return func(q *gcLeaderboardSetsQuery) {
+		if limit > 0 {
+			q.limit = limit
+		}
+	}
+}
+
+// WithGCLeaderboardSetsNextURL uses a next page URL directly.
+func WithGCLeaderboardSetsNextURL(next string) GCLeaderboardSetsOption {
+	return func(q *gcLeaderboardSetsQuery) {
+		if strings.TrimSpace(next) != "" {
+			q.nextURL = strings.TrimSpace(next)
+		}
+	}
+}
+
+func buildGCLeaderboardSetsQuery(query *gcLeaderboardSetsQuery) string {
+	values := url.Values{}
+	addLimit(values, query.limit)
+	return values.Encode()
+}
+
+// GameCenterLeaderboardSetLocalizationAttributes represents a Game Center leaderboard set localization resource.
+type GameCenterLeaderboardSetLocalizationAttributes struct {
+	Locale string `json:"locale"`
+	Name   string `json:"name,omitempty"`
+}
+
+// GameCenterLeaderboardSetLocalizationCreateAttributes describes attributes for creating a leaderboard set localization.
+type GameCenterLeaderboardSetLocalizationCreateAttributes struct {
+	Locale string `json:"locale"`
+	Name   string `json:"name,omitempty"`
+}
+
+// GameCenterLeaderboardSetLocalizationUpdateAttributes describes attributes for updating a leaderboard set localization.
+type GameCenterLeaderboardSetLocalizationUpdateAttributes struct {
+	Name *string `json:"name,omitempty"`
+}
+
+// GameCenterLeaderboardSetLocalizationRelationships describes relationships for leaderboard set localizations.
+type GameCenterLeaderboardSetLocalizationRelationships struct {
+	GameCenterLeaderboardSet *Relationship `json:"gameCenterLeaderboardSet"`
+}
+
+// GameCenterLeaderboardSetLocalizationCreateData is the data portion of a leaderboard set localization create request.
+type GameCenterLeaderboardSetLocalizationCreateData struct {
+	Type          ResourceType                                         `json:"type"`
+	Attributes    GameCenterLeaderboardSetLocalizationCreateAttributes `json:"attributes"`
+	Relationships *GameCenterLeaderboardSetLocalizationRelationships   `json:"relationships,omitempty"`
+}
+
+// GameCenterLeaderboardSetLocalizationCreateRequest is a request to create a leaderboard set localization.
+type GameCenterLeaderboardSetLocalizationCreateRequest struct {
+	Data GameCenterLeaderboardSetLocalizationCreateData `json:"data"`
+}
+
+// GameCenterLeaderboardSetLocalizationUpdateData is the data portion of a leaderboard set localization update request.
+type GameCenterLeaderboardSetLocalizationUpdateData struct {
+	Type       ResourceType                                          `json:"type"`
+	ID         string                                                `json:"id"`
+	Attributes *GameCenterLeaderboardSetLocalizationUpdateAttributes `json:"attributes,omitempty"`
+}
+
+// GameCenterLeaderboardSetLocalizationUpdateRequest is a request to update a leaderboard set localization.
+type GameCenterLeaderboardSetLocalizationUpdateRequest struct {
+	Data GameCenterLeaderboardSetLocalizationUpdateData `json:"data"`
+}
+
+// GameCenterLeaderboardSetLocalizationsResponse is the response from leaderboard set localization list endpoints.
+type GameCenterLeaderboardSetLocalizationsResponse = Response[GameCenterLeaderboardSetLocalizationAttributes]
+
+// GameCenterLeaderboardSetLocalizationResponse is the response from leaderboard set localization detail endpoints.
+type GameCenterLeaderboardSetLocalizationResponse = SingleResponse[GameCenterLeaderboardSetLocalizationAttributes]
+
+// GameCenterLeaderboardSetLocalizationDeleteResult represents CLI output for leaderboard set localization deletions.
+type GameCenterLeaderboardSetLocalizationDeleteResult struct {
+	ID      string `json:"id"`
+	Deleted bool   `json:"deleted"`
+}
+
+// GCLeaderboardSetLocalizationsOption is a functional option for GetGameCenterLeaderboardSetLocalizations.
+type GCLeaderboardSetLocalizationsOption func(*gcLeaderboardSetLocalizationsQuery)
+
+type gcLeaderboardSetLocalizationsQuery struct {
+	listQuery
+}
+
+// WithGCLeaderboardSetLocalizationsLimit sets the max number of leaderboard set localizations to return.
+func WithGCLeaderboardSetLocalizationsLimit(limit int) GCLeaderboardSetLocalizationsOption {
+	return func(q *gcLeaderboardSetLocalizationsQuery) {
+		if limit > 0 {
+			q.limit = limit
+		}
+	}
+}
+
+// WithGCLeaderboardSetLocalizationsNextURL uses a next page URL directly.
+func WithGCLeaderboardSetLocalizationsNextURL(next string) GCLeaderboardSetLocalizationsOption {
+	return func(q *gcLeaderboardSetLocalizationsQuery) {
+		if strings.TrimSpace(next) != "" {
+			q.nextURL = strings.TrimSpace(next)
+		}
+	}
+}
+
+func buildGCLeaderboardSetLocalizationsQuery(query *gcLeaderboardSetLocalizationsQuery) string {
+	values := url.Values{}
+	addLimit(values, query.limit)
+	return values.Encode()
+}
+
+// GameCenterLeaderboardSetReleaseAttributes represents a Game Center leaderboard set release resource.
+type GameCenterLeaderboardSetReleaseAttributes struct {
+	Live bool `json:"live"`
+}
+
+// GameCenterLeaderboardSetReleaseRelationships describes relationships for leaderboard set releases.
+type GameCenterLeaderboardSetReleaseRelationships struct {
+	GameCenterDetail         *Relationship `json:"gameCenterDetail"`
+	GameCenterLeaderboardSet *Relationship `json:"gameCenterLeaderboardSet"`
+}
+
+// GameCenterLeaderboardSetReleaseCreateData is the data portion of a leaderboard set release create request.
+type GameCenterLeaderboardSetReleaseCreateData struct {
+	Type          ResourceType                                  `json:"type"`
+	Relationships *GameCenterLeaderboardSetReleaseRelationships `json:"relationships"`
+}
+
+// GameCenterLeaderboardSetReleaseCreateRequest is a request to create a leaderboard set release.
+type GameCenterLeaderboardSetReleaseCreateRequest struct {
+	Data GameCenterLeaderboardSetReleaseCreateData `json:"data"`
+}
+
+// GameCenterLeaderboardSetReleasesResponse is the response from leaderboard set release list endpoints.
+type GameCenterLeaderboardSetReleasesResponse = Response[GameCenterLeaderboardSetReleaseAttributes]
+
+// GameCenterLeaderboardSetReleaseResponse is the response from leaderboard set release detail endpoints.
+type GameCenterLeaderboardSetReleaseResponse = SingleResponse[GameCenterLeaderboardSetReleaseAttributes]
+
+// GameCenterLeaderboardSetReleaseDeleteResult represents CLI output for leaderboard set release deletions.
+type GameCenterLeaderboardSetReleaseDeleteResult struct {
+	ID      string `json:"id"`
+	Deleted bool   `json:"deleted"`
+}
+
+// GCLeaderboardSetReleasesOption is a functional option for GetGameCenterLeaderboardSetReleases.
+type GCLeaderboardSetReleasesOption func(*gcLeaderboardSetReleasesQuery)
+
+type gcLeaderboardSetReleasesQuery struct {
+	listQuery
+}
+
+// WithGCLeaderboardSetReleasesLimit sets the max number of leaderboard set releases to return.
+func WithGCLeaderboardSetReleasesLimit(limit int) GCLeaderboardSetReleasesOption {
+	return func(q *gcLeaderboardSetReleasesQuery) {
+		if limit > 0 {
+			q.limit = limit
+		}
+	}
+}
+
+// WithGCLeaderboardSetReleasesNextURL uses a next page URL directly.
+func WithGCLeaderboardSetReleasesNextURL(next string) GCLeaderboardSetReleasesOption {
+	return func(q *gcLeaderboardSetReleasesQuery) {
+		if strings.TrimSpace(next) != "" {
+			q.nextURL = strings.TrimSpace(next)
+		}
+	}
+}
+
+func buildGCLeaderboardSetReleasesQuery(query *gcLeaderboardSetReleasesQuery) string {
+	values := url.Values{}
+	addLimit(values, query.limit)
+	return values.Encode()
+}
+
+// GameCenterLeaderboardSetMembersUpdateRequest is a request to replace leaderboard set members.
+type GameCenterLeaderboardSetMembersUpdateRequest struct {
+	Data []RelationshipData `json:"data"`
+}
+
+// GameCenterLeaderboardSetMembersUpdateResult represents CLI output for member updates.
+type GameCenterLeaderboardSetMembersUpdateResult struct {
+	SetID       string   `json:"setId"`
+	MemberCount int      `json:"memberCount"`
+	MemberIDs   []string `json:"memberIds"`
+	Updated     bool     `json:"updated"`
+}
+
+// GCLeaderboardSetMembersOption is a functional option for GetGameCenterLeaderboardSetMembers.
+type GCLeaderboardSetMembersOption func(*gcLeaderboardSetMembersQuery)
+
+type gcLeaderboardSetMembersQuery struct {
+	listQuery
+}
+
+// WithGCLeaderboardSetMembersLimit sets the max number of members to return.
+func WithGCLeaderboardSetMembersLimit(limit int) GCLeaderboardSetMembersOption {
+	return func(q *gcLeaderboardSetMembersQuery) {
+		if limit > 0 {
+			q.limit = limit
+		}
+	}
+}
+
+// WithGCLeaderboardSetMembersNextURL uses a next page URL directly.
+func WithGCLeaderboardSetMembersNextURL(next string) GCLeaderboardSetMembersOption {
+	return func(q *gcLeaderboardSetMembersQuery) {
+		if strings.TrimSpace(next) != "" {
+			q.nextURL = strings.TrimSpace(next)
+		}
+	}
+}
+
+func buildGCLeaderboardSetMembersQuery(query *gcLeaderboardSetMembersQuery) string {
+	values := url.Values{}
+	addLimit(values, query.limit)
+	return values.Encode()
+}
+
+// GameCenterLeaderboardSetImageAttributes represents a Game Center leaderboard set image resource.
+type GameCenterLeaderboardSetImageAttributes struct {
+	FileSize           int64               `json:"fileSize"`
+	FileName           string              `json:"fileName"`
+	ImageAsset         *ImageAsset         `json:"imageAsset,omitempty"`
+	UploadOperations   []UploadOperation   `json:"uploadOperations,omitempty"`
+	AssetDeliveryState *AssetDeliveryState `json:"assetDeliveryState,omitempty"`
+}
+
+// GameCenterLeaderboardSetImageCreateAttributes describes attributes for reserving an image upload.
+type GameCenterLeaderboardSetImageCreateAttributes struct {
+	FileSize int64  `json:"fileSize"`
+	FileName string `json:"fileName"`
+}
+
+// GameCenterLeaderboardSetImageUpdateAttributes describes attributes for committing an image upload.
+type GameCenterLeaderboardSetImageUpdateAttributes struct {
+	Uploaded *bool `json:"uploaded,omitempty"`
+}
+
+// GameCenterLeaderboardSetImageRelationships describes relationships for leaderboard set images.
+type GameCenterLeaderboardSetImageRelationships struct {
+	GameCenterLeaderboardSetLocalization *Relationship `json:"gameCenterLeaderboardSetLocalization"`
+}
+
+// GameCenterLeaderboardSetImageCreateData is the data portion of an image create (reserve) request.
+type GameCenterLeaderboardSetImageCreateData struct {
+	Type          ResourceType                                  `json:"type"`
+	Attributes    GameCenterLeaderboardSetImageCreateAttributes `json:"attributes"`
+	Relationships *GameCenterLeaderboardSetImageRelationships   `json:"relationships"`
+}
+
+// GameCenterLeaderboardSetImageCreateRequest is a request to reserve an image upload.
+type GameCenterLeaderboardSetImageCreateRequest struct {
+	Data GameCenterLeaderboardSetImageCreateData `json:"data"`
+}
+
+// GameCenterLeaderboardSetImageUpdateData is the data portion of an image update (commit) request.
+type GameCenterLeaderboardSetImageUpdateData struct {
+	Type       ResourceType                                   `json:"type"`
+	ID         string                                         `json:"id"`
+	Attributes *GameCenterLeaderboardSetImageUpdateAttributes `json:"attributes,omitempty"`
+}
+
+// GameCenterLeaderboardSetImageUpdateRequest is a request to update a leaderboard set image.
+type GameCenterLeaderboardSetImageUpdateRequest struct {
+	Data GameCenterLeaderboardSetImageUpdateData `json:"data"`
+}
+
+// GameCenterLeaderboardSetImageResponse is the response from leaderboard set image detail endpoints.
+type GameCenterLeaderboardSetImageResponse = SingleResponse[GameCenterLeaderboardSetImageAttributes]
+
+// GameCenterLeaderboardSetImageDeleteResult represents CLI output for leaderboard set image deletions.
+type GameCenterLeaderboardSetImageDeleteResult struct {
+	ID      string `json:"id"`
+	Deleted bool   `json:"deleted"`
+}
+
+// GameCenterLeaderboardSetImageUploadResult represents CLI output for leaderboard set image uploads.
+type GameCenterLeaderboardSetImageUploadResult struct {
+	ID                 string `json:"id"`
+	LocalizationID     string `json:"localizationId"`
+	FileName           string `json:"fileName"`
+	FileSize           int64  `json:"fileSize"`
+	AssetDeliveryState string `json:"assetDeliveryState,omitempty"`
+	Uploaded           bool   `json:"uploaded"`
+}
